% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{stablehlopt_bin}
\alias{stablehlopt_bin}
\title{Get the path to the stablehlo-opt binary}
\usage{
stablehlopt_bin(install = TRUE)
}
\arguments{
\item{install}{Logical, whether to install the binary if not found. Default is TRUE.}
}
\value{
The file path to the stablehlo-opt binary.
}
\description{
This function returns the path to the stablehlo-opt binary. If the binary
is not found and \code{install} is TRUE, it will attempt to download and install
the binary.
}
